Separate out python south plugin interfacing logic to a separate library which can be loaded on demand when required
