+++ CMakeLists.txt
@@ -13,6 +13,7 @@ add_subdirectory(C/plugins/storage/postgres)
 add_subdirectory(C/plugins/storage/sqlite)
 add_subdirectory(C/plugins/storage/sqlitememory)
 add_subdirectory(C/services/south)
+add_subdirectory(C/services/south-plugin-interfaces/python)
 add_subdirectory(C/tasks/north)
 add_subdirectory(C/plugins/utils)
 add_subdirectory(C/plugins/north/PI_Server_V2)
